+++ src/Screens.tsx
@@ -12,6 +12,7 @@ import { InvoiceSelectActionScreen } from './screens/InvoiceSelectActionScreen';
 import { InvoiceSelectUserScreen } from './screens/InvoiceSelectUserScreen';
 import { InvoicesToApproveScreen } from './screens/InvoicesToApproveScreen';
 import { InvoiceTimelineScreen } from './screens/InvoiceTimelineScreen';
+import { ScanCameraScreen } from './screens/ScanCameraScreen';
 import { ScanPreviewScreen } from './screens/ScanPreviewScreen';
 import { SettingsScreen } from './screens/SettingsScreen';
 import { SwitchEnvironmentScreen } from './screens/SwitchEnvironmentScreen';
@@ -105,7 +106,18 @@ const Screens: React.FC = () => {
             name="ScanPreviewScreen"
             component={ScanPreviewScreen}
             options={{
-              title: t('scan.screen_title') ?? '',
+              headerBackTitle: t('dashboard.title') ?? '',
+              headerShown: false,
+              title: t('scan.preview_title') ?? '',
+            }}
+          />
+          <Stack.Screen
+            name="ScanCameraScreen"
+            component={ScanCameraScreen}
+            options={{
+              headerBackTitle: t('scan.preview_title') ?? '',
+              headerShown: false,
+              title: t('scan.camera_title') ?? '',
             }}
           />
         </Stack.Navigator>